feat(affiliate): enhance referral code verification logic and UI components
